+++ src/_modules/function.ts
@@ -4,10 +4,12 @@ import bind from '../function/bind'
 import bindAll from '../function/bindAll'
 import call from '../function/call'
 import compose from '../function/compose'
+import debounce from '../function/debounce'
 import delay from '../function/delay'
 import fval from '../function/fval'
 import once from '../function/once'
 import partial from '../function/partial'
 import tap from '../function/tap'
+import throttle from '../function/throttle'
 
-export {after,alt,bind,bindAll,call,compose,delay,fval,once,partial,tap}
+export {after,alt,bind,bindAll,call,compose,debounce,delay,fval,once,partial,tap,throttle}