+++ src/collection/reduce.ts
@@ -25,6 +25,16 @@ import each from "./each";
  * @param [initialValue] 第一次调用 callback函数时的第一个参数的值
  * @returns 汇总值
  */
+function reduce<T,U>(
+  collection: Collection<T>,
+  callback: (
+    accumulator: U,
+    value: T,
+    key: UnknownMapKey,
+    collection: Collection<T>
+  ) => U,
+  initialValue: U
+): U 
 function reduce<T>(
   collection: Collection<T>,
   callback: (